Identify written contents variable and flag names (#136)
* name item_flags2[3] "has_written_content"

item_flags2[3] is the flag that indicates that a quire or scroll has one or more itemimprovement_writingst improvements on it.  It's used by workshops to determine if an object might meet the [HAS_WRITING_IMPROVEMENT] requirement for reactions like BIND_BOOK.  (There's also a further check to see if there's an actual improvement in the improvements list.)

* name itemimprovement_writingst::anon_1

itemimprovement_writingst::anon_1 contains the writings in a quire, scroll, or book.  The elements are IDs into df.global.world.written_contents.all.  Naming this variable will also fix the inconvenient name collision with the anon_1 in the base class.