Name 2 historical_kills fields & fix undead flags
Just as `killed_site` is a list of site IDs where unimportant things
were killed, `killed_underground_region` and `killed_region` are lists
of underground region and region IDs.

The `skeletal` bit in `killed_undead` was a relic of old versions of the
game where creatures could be skeletons or zombies. Zombies live on in
the current night creature ark but skeletons are no more.