Fix a couple ui_sidebar_menus x64 alignment issues
show_misc and barracks_squad_cursor have just one byte of padding between them
on x64, so they can't be inside of the barracks and unit_skills compounds. It's
possible that more of these fields are outside compounds, but this is the
minimal change needed to align barracks and command_line (below).

Also fixed an issue with ui_sidebar_menus.units that was causing accesses to
job_details fields to crash.