Identify the adventurer's commander's ID
In the "Create Your Character" screen of adventurer creation, you can
choose "Locally Important" or "Peasant". "Locally Important" means the
chosen site has a captain of the guard position (or equivalent).
`commander_id` is the index of the position assignment of the captain of
the guard for the site's group. If position is vacant, the "Locally
Important" option is available, but the adventurer is still a peasant
and `commander_id` is -1.