Add hex values to df enums
Seeing enum values in header makes it easier to match those values to
dissambler numbers. This helps me with both gdb and IDA. IDA has descent
suppot for offsets and constands but it requires manual setup. To help
matching values manually I like to browser headers in my source editor
because it is the fastest tool to find types and variables.