Renamed "read_books" to "known_written_contents"
More accurate as `read_books` implies that this refers to the artifact or item ID of books read, which is not the case, as this is a list of written_content IDs. More importantly, the ID of art form derivations including choreography and musical compositions known to the historical figure are also listed here, as these derivations are stored in the form of a `written_content` despite being unrelated to books.