fix enum definitions of hauler_type
while implementing the --hauler-type param for the prioritize script, I
noticed that the job->item_subtype field, which is tagged as being a
hauler_type enum, did not match up with the type of item being hauled. I
fixed this enum based on observation of which enum values resulted in
which types of items were hauled.

I was unabled to verify the "Bin" enum. I could not get any hauling jobs
of this type to be generated in my fort, even when explicitly building
bins and enabling bins for stockpiles.