[SDL2 port] codegen: generate friend declarations for classes that contain members with vtables (#1)
* Update symbols.xml

* add vtable addresses for widgets

* add instance variables to `widget`

cuz i somehow missed that widgets have instance variables, d'oh

* align gamest in sdl2 beta

* correctly align `gamest`

Co-Authored-By: Quietust <1005195+quietust@users.noreply.github.com>

* update `enablerst`

padding was used for the mutex and condition variables because we don't support these types in df-structures at present

* update `viewscreenst` for sdl2 beta

* sdl2 updates: `renderer` and `renderer_2d_base`

* sdl2: update `unitst`

* Generate friend declarations and forward-declarations for structs that contain class-type members

This is needed because class-type constructors (i.e. constructors for types with vmethods) are protected.

* Move struct field reference detection to a new preprocessing stage

This ensures that fields of classes requiring `friend` declarations are
registered properly, regardless of their names. Previously, these references
were only generated properly if the containing type came before the field type
alphabetically.

* Split main_interface_settings into a standalone type

Needed to allow widget_tabs to forward-declare it as a friend

---------

Co-authored-by: Kelly Kinkade <kelly.lynn.martin@gmail.com>
Co-authored-by: Quietust <1005195+quietust@users.noreply.github.com>
Co-authored-by: Kelly Kinkade <kmartin@pyrzqxgl.org>